+++ node/service/src/rpc.rs
@@ -56,7 +56,7 @@ where
 	use substrate_frame_rpc_system::{System, SystemApiServer};
 
 	// Unfinished RPCs
-	use pallet_messages_rpc::{MessagesApiServer, MessagesHandler};
+	use pallet_messages_rpc::rpc::{MessagesApiServer, MessagesHandler};
 	use pallet_msa_rpc::{MsaApiServer, MsaHandler};
 	use pallet_schemas_rpc::{SchemasApiServer, SchemasHandler};
